* ajuste modelo producto para que tome los valores decimales para el descuento

+++ persistencia/src/main/java/co/edu/uniquindio/proyecto/entidades/Producto.java
@@ -43,7 +43,7 @@ public class Producto implements Serializable {
 
     @Column(nullable = false)
     @Positive
-    private Integer descuento;
+    private Float descuento;
 
     @OneToMany(mappedBy = "producto")
     private List<Comentario> listaComentarios;
